VANIR_BUILD: fix false negatives in envsetup possible when comparing empty strings

+++ envsetup.sh
@@ -823,7 +823,7 @@ function eat()
             done
             echo "Device Found.."
         fi
-    if (adb shell getprop ro.cm.device | grep -q "$CM_BUILD");
+    if ([ $CM_BUILD ] && adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || ([ $VANIR_BUILD ] && adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD");
     then
         # if adbd isn't root we can't write to /cache/recovery/
         adb root
@@ -845,7 +845,7 @@ EOF
     fi
     return $?
     else
-        echo "The connected device does not appear to be $CM_BUILD, run away!"
+        echo "The connected device does not appear to be ${CM_BUILD}${VANIR_BUILD}, run away!"
     fi
 }
 
@@ -1860,7 +1860,7 @@ function installboot()
     sleep 1
     adb wait-for-online shell mount /system 2>&1 > /dev/null
     adb wait-for-online remount
-    if (adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || (adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD");
+    if ([ $CM_BUILD ] && adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || ([ $VANIR_BUILD ] && adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD");
     then
         adb push $OUT/boot.img /cache/
         for i in $OUT/system/lib/modules/*;
@@ -1905,7 +1905,7 @@ function installrecovery()
     sleep 1
     adb wait-for-online shell mount /system 2>&1 >> /dev/null
     adb wait-for-online remount
-    if (adb shell getprop ro.cm.device | grep -q "$CM_BUILD") ||(adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD");
+    if ([ $CM_BUILD ] && adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || ([ $VANIR_BUILD ] && adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD");
     then
         adb push $OUT/recovery.img /cache/
         adb shell dd if=/cache/recovery.img of=$PARTITION
@@ -2109,7 +2109,7 @@ function dopush()
         echo "Device Found."
     fi
 
-    if (adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || (adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD") || [ "$FORCE_PUSH" == "true" ];
+    if ([ $CM_BUILD ] && adb shell getprop ro.cm.device | grep -q "$CM_BUILD") || ([ $VANIR_BUILD ] && adb shell getprop ro.vanir.device | grep -q "$VANIR_BUILD") || [ "$FORCE_PUSH" == "true" ];
     then
     # retrieve IP and PORT info if we're using a TCP connection
     TCPIPPORT=$(adb devices | egrep '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+:[0-9]+[^0-9]+' \
@@ -2220,7 +2220,7 @@ EOF
     rm -f $OUT/.log
     return 0
     else
-        echo "The connected device does not appear to be $CM_BUILD, run away!"
+        echo "The connected device does not appear to be ${CM_BUILD}${VANIR_BUILD}, run away!"
     fi
 }
